refactor: 파일의 역할 범주에 따른 패키지로 이동
