+++ src/main/java/dto/HttpResponse.java
@@ -1,7 +1,8 @@
-package utility;
+package dto;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import utility.HttpStatusCode;
 
 import java.io.DataOutputStream;
 import java.io.File;
